requirements.txt
# Core web framework
flask==3.0.2

# LangChain ecosystem (PINNED for compatibility)
langchain==0.3.27
langchain-core==0.3.75
langgraph==0.2.55

# LLM integrations
langchain-groq==0.3.6
langchain-google-genai==2.0.10

# NLP utilities
nltk==3.8.1
textstat==0.7.4

# Environment management
python-dotenv==1.0.1

# Google SDK (for model listing/debugging)
google-generativeai==0.5.4
